Extract TypeManager's pure conversion logic into a net8.0 assembly (#2276)
TypeManager (Glue.csproj, net8.0-windows) is deeply coupled to Glue's plugin/asset-type infrastructure for its reflection-based type resolution (GetTypeFromString), so that half stays put - but its string/value conversion methods (GetCommonTypeName, GetDefaultForType, TryConvertStringValue, TryCastValue, Parse, GetFriendlyGenericName, GetElementType) have no such dependency.

Moved that subset to a new TypeConversion class in GlueCommon (already net8.0, no WPF). TypeManager's own methods now just forward to it, so none of Glue.csproj's ~15 call sites change. Added GlueCommonUnitTests (net8.0 xUnit) with 32 tests covering the moved logic, wired it into 'Glue with All.sln', and added a ubuntu-latest CI job that runs it - the first Glue-derived test suite that builds and runs on Linux.

Finding for #2276: TypeManager's type-resolution half (GetTypeFromString) is not incidentally coupled to DialogService/AvailableAssetTypes/Plugins.PluginManager the way the issue speculated - it genuinely calls into all three (a plugin-type scan, an AssetTypeInfo lookup, and a WPF-backed error dialog), so moving it needs those seams built first, not just a file move.

Manual test: not needed - covered by the new unit tests plus Glue.csproj/GlueCommon.csproj compiling; verified locally with dotnet build/test (this box has the SDK) before pushing: GlueCommon, Glue.csproj, and the full 'Glue with All.sln' build clean, the 957-test fast GlueUnitTests suite still passes, and the new GlueCommonUnitTests project's 32 tests pass both standalone and via the solution.

Part of #2276.
