+++ FRBDK/Glue/Glue with All.sln
@@ -95,6 +95,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SkiaGum", "..\..\..\Gum\Run
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tests", "Tests", "{3CB2E620-1431-4DEA-AD58-E844D5DDCC60}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GlueCommonUnitTests", "Tests\GlueCommonUnitTests\GlueCommonUnitTests.csproj", "{83708235-A2F6-4DEF-8E0D-EAF4AA002944}"
+EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GlueUnitTests", "Tests\GlueUnitTests\GlueUnitTests.csproj", "{ABCCA529-A8A3-44E4-B2E8-C69EA3B00644}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Gum.Analyzers", "..\..\..\Gum\Tools\Gum.Analyzers\Gum.Analyzers.csproj", "{8A17AC92-C2A1-35C1-910E-858AD7715352}"
@@ -231,6 +233,10 @@ Global
 		{3AC20D6E-C39F-91E6-8942-F44C8430814A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{3AC20D6E-C39F-91E6-8942-F44C8430814A}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{3AC20D6E-C39F-91E6-8942-F44C8430814A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{83708235-A2F6-4DEF-8E0D-EAF4AA002944}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{83708235-A2F6-4DEF-8E0D-EAF4AA002944}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{83708235-A2F6-4DEF-8E0D-EAF4AA002944}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{83708235-A2F6-4DEF-8E0D-EAF4AA002944}.Release|Any CPU.Build.0 = Release|Any CPU
 		{ABCCA529-A8A3-44E4-B2E8-C69EA3B00644}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{ABCCA529-A8A3-44E4-B2E8-C69EA3B00644}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{ABCCA529-A8A3-44E4-B2E8-C69EA3B00644}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -285,6 +291,7 @@ Global
 		{74D65A46-28C0-49FB-938B-6ED83C06DC73} = {60818A85-E5BC-4B05-97F9-24402B4B0CCE}
 		{1ACC516A-D3D4-B8D5-B98B-87743A46466C} = {07985C4C-B3CF-432C-8A27-36FDD006231F}
 		{3AC20D6E-C39F-91E6-8942-F44C8430814A} = {07985C4C-B3CF-432C-8A27-36FDD006231F}
+		{83708235-A2F6-4DEF-8E0D-EAF4AA002944} = {3CB2E620-1431-4DEA-AD58-E844D5DDCC60}
 		{ABCCA529-A8A3-44E4-B2E8-C69EA3B00644} = {3CB2E620-1431-4DEA-AD58-E844D5DDCC60}
 		{8A17AC92-C2A1-35C1-910E-858AD7715352} = {07985C4C-B3CF-432C-8A27-36FDD006231F}
 		{317E351A-7599-90A4-03FD-991745D21AC3} = {07985C4C-B3CF-432C-8A27-36FDD006231F}
